long/int fehler fix

+++ src/main/java/htw/webtech/WT_todo/business/service/TodoService.java
@@ -74,7 +74,7 @@ public class TodoService {
     }
 
     @Transactional
-    public long deleteCompleted() {
+    public int deleteCompleted() {
         return todoRepository.deleteByDoneTrue();
     }
